+++ generator/sql_types.cpp
@@ -479,12 +479,12 @@ void StatementGenerator::StrBuildJSONElement(RandomGenerator &rg, std::string &r
 		case 9:
 		case 10: { //string
 			std::uniform_int_distribution<int> slen(0, 10);
-			std::uniform_int_distribution<int> chars(32, 128);
+			std::uniform_int_distribution<uint8_t> chars(32, 128);
 			const int nlen = slen(rg.gen);
 
 			ret += '"';
 			for (int i = 0 ; i < nlen ; i++) {
-				const int nchar = chars(rg.gen);
+				const uint8_t nchar = chars(rg.gen);
 
 				switch (nchar) {
 				case 128:
